feat: Added 404 page and input validation for join links
- Any URLs not in the routes list are now automatically redirected to the 404 page.
- When joining a lobby using a link, the app now checks that the URL contains a valid pin number format before continuing, or redirects the user to the 404 page.
- Removed the old, redundant NicknamePage.js and changed Ably.js to NicknamePage.js to better reflect its function.
- Organised imports on each page file.
