+++ src/pages/Homepage.js
@@ -1,8 +1,13 @@
+// React
 import React, { useState } from 'react';
-import '../App.css';
+import { NavLink } from 'react-router-dom';
+
+// Components
 import Button from '../components/Button.js'
 import Input from '../components/Input.js'
-import { NavLink } from 'react-router-dom';
+
+// Styles
+import '../App.css';
 
 const Homepage = () => {
     const [gamePin, setGamePin] = useState("");
@@ -17,7 +22,7 @@ const Homepage = () => {
                     maxLength={5}
                     onChange={ (e) => setGamePin(e.target.value) }
                 />
-                <NavLink to={`/Lobby/${gamePin}`} state={{ isHost: false, joinPin: gamePin }}>
+                <NavLink to={`/Lobby/link/${gamePin}`} state={{ isHost: false, joinPin: gamePin }}>
                     <Button 
                         name="JOIN" //check if the provided Game Pin is at least 5 characters long, only contains numbers and isn't made up of only whitespace
                         disabled={ (gamePin.length === 5) && (/^[0-9\b]+$/.test(gamePin)) && (/\S/.test(gamePin)) ? false : true }
